Create Static and Class Methods (Alternative Constructor)
In this exercise I create a 'Luxury watch' class which contains instance and class attributes as well as instance, class and static methods. I use a class attribute to store the number of watches created, and a class method to access and return the number. 
I also use a class method as an alternative constructor which allows for additional features for the watch e.g. engraving. I use a static method to validate the engraving text.